third_party/WebKit/LayoutTests/fast/dom/node-legacy-event-listener.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Test addEventListener() and removeEventListener() fail silently if arguments are missing.");

shouldThrow('document.addEventListener("foo")');
shouldThrow('document.removeEventListener("bar")');

shouldThrow('document.addEventListener()');
shouldThrow('document.removeEventListener()');
</script>
</body>
</html>
